Fix other tutorials

+++ examples/dmm/polyphonic_data_loader.py
@@ -118,9 +118,9 @@ def get_mini_batch(mini_batch_indices, sequences, seq_lengths, cuda=False):
     mini_batch_mask = get_mini_batch_mask(mini_batch, sorted_seq_lengths)
 
     # wrap in PyTorch Tensors
-    mini_batch = torch.tensor(mini_batch)
-    mini_batch_reversed = torch.tensor(mini_batch_reversed)
-    mini_batch_mask = torch.tensor(mini_batch_mask)
+    mini_batch = torch.tensor(mini_batch, dtype=torch.float32)
+    mini_batch_reversed = torch.tensor(mini_batch_reversed, dtype=torch.float32)
+    mini_batch_mask = torch.tensor(mini_batch_mask, dtype=torch.float32)
 
     # cuda() here because need to cuda() before packing
     if cuda:
